// webapp/src/app/model/user.ts
export class User {
    id: any;
    firstName: string = "";
    lastName: string = "";
    email: string = "";
    password: string = "";
    city: string = "";
    country: string = "";
    occupation: string = "";
    companyInfo: string = "";
    phoneNumber: string = "";
    role: string = "";
}